requirements.txt
requests
click
lxml
simple_model
ruamel.yaml>=0.12.12
